Moving _parallel_process.py from magicc to common adapter utils folder
